requirements.txt
numpy
pandas
python-dotenv
datetime
openai
anthropic
google
genai